Tries to avoid multiple instanciations of panels in legends editors. See #484. There are at least two instanciations provoked by the code in SimpleStyleEditor : one when calling initialize, one when calling setLegend. In fact, setLegend is responsible to the calling of initializeLegendFields, which is the method that creates most of the UIs. When calling initialize, we make a silent call to setLegend... and then we call setLegend again.
To fix that, I add a initialize(LegendContext lc, Legend leg) to the APi
of ILegendPanel. calling it, we don't need the external call to
setLegend : we admit that we are giving the needed legend to the panel
from the beginning. The original initialize is still here and uses the
new one.